+++ src/Meta_Auto_Code.cpp
@@ -346,7 +346,7 @@ struct Meta_Auto_Code::impl : public Meta_Auto_Code_command_handler
 
    bool value_will_be_provided( const string& field_name );
 
-   void validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors );
+   void validate( uint64_t state, bool is_internal, validation_error_container* p_validation_errors );
    void validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors );
 
    void after_fetch( );
@@ -637,7 +637,8 @@ bool Meta_Auto_Code::impl::value_will_be_provided( const string& field_name )
    return false;
 }
 
-void Meta_Auto_Code::impl::validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors )
+void Meta_Auto_Code::impl::validate(
+ uint64_t state, bool is_internal, validation_error_container* p_validation_errors )
 {
    ( void )state;
    ( void )is_internal;
@@ -712,7 +713,8 @@ void Meta_Auto_Code::impl::validate( unsigned state, bool is_internal, validatio
    // [<finish validate>]
 }
 
-void Meta_Auto_Code::impl::validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors )
+void Meta_Auto_Code::impl::validate_set_fields(
+ set< string >& fields_set, validation_error_container* p_validation_errors )
 {
    ( void )fields_set;
 
@@ -1038,7 +1040,7 @@ void Meta_Auto_Code::clear( )
    p_impl->clear( );
 }
 
-void Meta_Auto_Code::validate( unsigned state, bool is_internal )
+void Meta_Auto_Code::validate( uint64_t state, bool is_internal )
 {
    p_impl->validate( state, is_internal, &validation_errors );
 }
