+++ src/Meta_Class.cpp
@@ -2238,7 +2238,7 @@ struct Meta_Class::impl : public Meta_Class_command_handler
 
    bool value_will_be_provided( const string& field_name );
 
-   void validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors );
+   void validate( uint64_t state, bool is_internal, validation_error_container* p_validation_errors );
    void validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors );
 
    void after_fetch( );
@@ -4398,7 +4398,8 @@ bool Meta_Class::impl::value_will_be_provided( const string& field_name )
    return false;
 }
 
-void Meta_Class::impl::validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors )
+void Meta_Class::impl::validate(
+ uint64_t state, bool is_internal, validation_error_container* p_validation_errors )
 {
    ( void )state;
    ( void )is_internal;
@@ -4511,7 +4512,8 @@ void Meta_Class::impl::validate( unsigned state, bool is_internal, validation_er
    // [<finish validate>]
 }
 
-void Meta_Class::impl::validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors )
+void Meta_Class::impl::validate_set_fields(
+ set< string >& fields_set, validation_error_container* p_validation_errors )
 {
    ( void )fields_set;
 
@@ -5981,7 +5983,7 @@ void Meta_Class::clear( )
    p_impl->clear( );
 }
 
-void Meta_Class::validate( unsigned state, bool is_internal )
+void Meta_Class::validate( uint64_t state, bool is_internal )
 {
    p_impl->validate( state, is_internal, &validation_errors );
 }
