+++ src/Meta_List.cpp
@@ -2206,7 +2206,7 @@ struct Meta_List::impl : public Meta_List_command_handler
 
    bool value_will_be_provided( const string& field_name );
 
-   void validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors );
+   void validate( uint64_t state, bool is_internal, validation_error_container* p_validation_errors );
    void validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors );
 
    void after_fetch( );
@@ -4474,7 +4474,8 @@ bool Meta_List::impl::value_will_be_provided( const string& field_name )
    return false;
 }
 
-void Meta_List::impl::validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors )
+void Meta_List::impl::validate(
+ uint64_t state, bool is_internal, validation_error_container* p_validation_errors )
 {
    ( void )state;
    ( void )is_internal;
@@ -4607,7 +4608,8 @@ void Meta_List::impl::validate( unsigned state, bool is_internal, validation_err
    // [<finish validate>]
 }
 
-void Meta_List::impl::validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors )
+void Meta_List::impl::validate_set_fields(
+ set< string >& fields_set, validation_error_container* p_validation_errors )
 {
    ( void )fields_set;
 
@@ -5672,7 +5674,7 @@ void Meta_List::clear( )
    p_impl->clear( );
 }
 
-void Meta_List::validate( unsigned state, bool is_internal )
+void Meta_List::validate( uint64_t state, bool is_internal )
 {
    p_impl->validate( state, is_internal, &validation_errors );
 }
