+++ src/Meta_List_Field_Type.cpp
@@ -776,7 +776,7 @@ struct Meta_List_Field_Type::impl : public Meta_List_Field_Type_command_handler
 
    bool value_will_be_provided( const string& field_name );
 
-   void validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors );
+   void validate( uint64_t state, bool is_internal, validation_error_container* p_validation_errors );
    void validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors );
 
    void after_fetch( );
@@ -1328,7 +1328,8 @@ bool Meta_List_Field_Type::impl::value_will_be_provided( const string& field_nam
    return false;
 }
 
-void Meta_List_Field_Type::impl::validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors )
+void Meta_List_Field_Type::impl::validate(
+ uint64_t state, bool is_internal, validation_error_container* p_validation_errors )
 {
    ( void )state;
    ( void )is_internal;
@@ -1365,7 +1366,8 @@ void Meta_List_Field_Type::impl::validate( unsigned state, bool is_internal, val
    // [<finish validate>]
 }
 
-void Meta_List_Field_Type::impl::validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors )
+void Meta_List_Field_Type::impl::validate_set_fields(
+ set< string >& fields_set, validation_error_container* p_validation_errors )
 {
    ( void )fields_set;
 
@@ -1886,7 +1888,7 @@ void Meta_List_Field_Type::clear( )
    p_impl->clear( );
 }
 
-void Meta_List_Field_Type::validate( unsigned state, bool is_internal )
+void Meta_List_Field_Type::validate( uint64_t state, bool is_internal )
 {
    p_impl->validate( state, is_internal, &validation_errors );
 }
