+++ src/Meta_Package.cpp
@@ -726,7 +726,7 @@ struct Meta_Package::impl : public Meta_Package_command_handler
 
    bool value_will_be_provided( const string& field_name );
 
-   void validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors );
+   void validate( uint64_t state, bool is_internal, validation_error_container* p_validation_errors );
    void validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors );
 
    void after_fetch( );
@@ -1962,7 +1962,8 @@ bool Meta_Package::impl::value_will_be_provided( const string& field_name )
    return false;
 }
 
-void Meta_Package::impl::validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors )
+void Meta_Package::impl::validate(
+ uint64_t state, bool is_internal, validation_error_container* p_validation_errors )
 {
    ( void )state;
    ( void )is_internal;
@@ -2014,7 +2015,8 @@ void Meta_Package::impl::validate( unsigned state, bool is_internal, validation_
    // [<finish validate>]
 }
 
-void Meta_Package::impl::validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors )
+void Meta_Package::impl::validate_set_fields(
+ set< string >& fields_set, validation_error_container* p_validation_errors )
 {
    ( void )fields_set;
 
@@ -2951,7 +2953,7 @@ void Meta_Package::clear( )
    p_impl->clear( );
 }
 
-void Meta_Package::validate( unsigned state, bool is_internal )
+void Meta_Package::validate( uint64_t state, bool is_internal )
 {
    p_impl->validate( state, is_internal, &validation_errors );
 }
