+++ src/Meta_Package.h
@@ -133,7 +133,7 @@ class META_PACKAGE_DECL_SPEC Meta_Package : public class_base
 
    void clear( );
 
-   void validate( unsigned state, bool is_internal );
+   void validate( uint64_t state, bool is_internal );
    void validate_set_fields( std::set< std::string >& fields_set );
 
    void after_fetch( );
