+++ src/Meta_Package_Option.cpp
@@ -1937,7 +1937,7 @@ struct Meta_Package_Option::impl : public Meta_Package_Option_command_handler
 
    bool value_will_be_provided( const string& field_name );
 
-   void validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors );
+   void validate( uint64_t state, bool is_internal, validation_error_container* p_validation_errors );
    void validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors );
 
    void after_fetch( );
@@ -3289,7 +3289,8 @@ bool Meta_Package_Option::impl::value_will_be_provided( const string& field_name
    return false;
 }
 
-void Meta_Package_Option::impl::validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors )
+void Meta_Package_Option::impl::validate(
+ uint64_t state, bool is_internal, validation_error_container* p_validation_errors )
 {
    ( void )state;
    ( void )is_internal;
@@ -3341,7 +3342,8 @@ void Meta_Package_Option::impl::validate( unsigned state, bool is_internal, vali
    // [<finish validate>]
 }
 
-void Meta_Package_Option::impl::validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors )
+void Meta_Package_Option::impl::validate_set_fields(
+ set< string >& fields_set, validation_error_container* p_validation_errors )
 {
    ( void )fields_set;
 
@@ -4317,7 +4319,7 @@ void Meta_Package_Option::clear( )
    p_impl->clear( );
 }
 
-void Meta_Package_Option::validate( unsigned state, bool is_internal )
+void Meta_Package_Option::validate( uint64_t state, bool is_internal )
 {
    p_impl->validate( state, is_internal, &validation_errors );
 }
