+++ src/Meta_Permission.cpp
@@ -893,7 +893,7 @@ struct Meta_Permission::impl : public Meta_Permission_command_handler
 
    bool value_will_be_provided( const string& field_name );
 
-   void validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors );
+   void validate( uint64_t state, bool is_internal, validation_error_container* p_validation_errors );
    void validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors );
 
    void after_fetch( );
@@ -1141,7 +1141,8 @@ bool Meta_Permission::impl::value_will_be_provided( const string& field_name )
    return false;
 }
 
-void Meta_Permission::impl::validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors )
+void Meta_Permission::impl::validate(
+ uint64_t state, bool is_internal, validation_error_container* p_validation_errors )
 {
    ( void )state;
    ( void )is_internal;
@@ -1183,7 +1184,8 @@ void Meta_Permission::impl::validate( unsigned state, bool is_internal, validati
    // [<finish validate>]
 }
 
-void Meta_Permission::impl::validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors )
+void Meta_Permission::impl::validate_set_fields(
+ set< string >& fields_set, validation_error_container* p_validation_errors )
 {
    ( void )fields_set;
 
@@ -1805,7 +1807,7 @@ void Meta_Permission::clear( )
    p_impl->clear( );
 }
 
-void Meta_Permission::validate( unsigned state, bool is_internal )
+void Meta_Permission::validate( uint64_t state, bool is_internal )
 {
    p_impl->validate( state, is_internal, &validation_errors );
 }
