+++ src/Meta_Procedure_Arg.cpp
@@ -678,7 +678,7 @@ struct Meta_Procedure_Arg::impl : public Meta_Procedure_Arg_command_handler
 
    bool value_will_be_provided( const string& field_name );
 
-   void validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors );
+   void validate( uint64_t state, bool is_internal, validation_error_container* p_validation_errors );
    void validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors );
 
    void after_fetch( );
@@ -1134,7 +1134,8 @@ bool Meta_Procedure_Arg::impl::value_will_be_provided( const string& field_name 
    return false;
 }
 
-void Meta_Procedure_Arg::impl::validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors )
+void Meta_Procedure_Arg::impl::validate(
+ uint64_t state, bool is_internal, validation_error_container* p_validation_errors )
 {
    ( void )state;
    ( void )is_internal;
@@ -1179,7 +1180,8 @@ void Meta_Procedure_Arg::impl::validate( unsigned state, bool is_internal, valid
    // [<finish validate>]
 }
 
-void Meta_Procedure_Arg::impl::validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors )
+void Meta_Procedure_Arg::impl::validate_set_fields(
+ set< string >& fields_set, validation_error_container* p_validation_errors )
 {
    ( void )fields_set;
 
@@ -1652,7 +1654,7 @@ void Meta_Procedure_Arg::clear( )
    p_impl->clear( );
 }
 
-void Meta_Procedure_Arg::validate( unsigned state, bool is_internal )
+void Meta_Procedure_Arg::validate( uint64_t state, bool is_internal )
 {
    p_impl->validate( state, is_internal, &validation_errors );
 }
