+++ src/Meta_Relationship.cpp
@@ -1358,7 +1358,7 @@ struct Meta_Relationship::impl : public Meta_Relationship_command_handler
 
    bool value_will_be_provided( const string& field_name );
 
-   void validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors );
+   void validate( uint64_t state, bool is_internal, validation_error_container* p_validation_errors );
    void validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors );
 
    void after_fetch( );
@@ -2050,7 +2050,8 @@ bool Meta_Relationship::impl::value_will_be_provided( const string& field_name )
    return false;
 }
 
-void Meta_Relationship::impl::validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors )
+void Meta_Relationship::impl::validate(
+ uint64_t state, bool is_internal, validation_error_container* p_validation_errors )
 {
    ( void )state;
    ( void )is_internal;
@@ -2143,7 +2144,8 @@ void Meta_Relationship::impl::validate( unsigned state, bool is_internal, valida
    // [<finish validate>]
 }
 
-void Meta_Relationship::impl::validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors )
+void Meta_Relationship::impl::validate_set_fields(
+ set< string >& fields_set, validation_error_container* p_validation_errors )
 {
    ( void )fields_set;
 
@@ -2984,7 +2986,7 @@ void Meta_Relationship::clear( )
    p_impl->clear( );
 }
 
-void Meta_Relationship::validate( unsigned state, bool is_internal )
+void Meta_Relationship::validate( uint64_t state, bool is_internal )
 {
    p_impl->validate( state, is_internal, &validation_errors );
 }
